Required fixes in torch_is_fid_score.py and remove not used code
